Replace the plot info display toggle with tooltips for the plot.
I believe this is better as it avoids the extra plot text shifting the layout of the table, and I think in most cases a user will just want to read the plot info, so having it expanded is unnecessary.
Also, working under the assumption that viewing plot info is a (relatively) rare thing to do (at least, I do it very rarely), the plot info is fetched by a synchronous call to the server when the tooltip is requested. This should make the show page load a bit faster, as I’ve removed all the plot info text from the page.