Clean up the browser code.
I’ve pulled out all the code related to the browser into a separate file, and rewritten parts of it in an attempt to make it more robust.
The same function is now used by both the browser and autocompleter.

All the browsing functionality should be solid under both systems (Unix/Windows) now.