Provide option under General to listen on ipv6
Listening on :: instead of 0.0.0.0 will provide the ability for
Sick Beard to provide its service to any connection to the host
CherryPy supports listening on :: fine, and there should be very
little effect except on platform that do not have ipv6 support at
all.

This option should probably remain being off by default just in
case sick beard is started on a host without ipv6 support and thus
cannot be reconfigured.

WEB_HOST is an option that exists but is not able to be set via
the web interface. This will take precedence over WEB_IPV6.