Use cherrypy tools.json_out() decorator to indicate JSON
CherryPy 3.2 includes a convenient decorator to indicate that an exposed
method should output JSON. Switch to using that instead of json.dumps
where possible to make it more clear what the intention of the method
is.

Unfortunately, the setStatus method has a parameter "direct" which
indicates whether it should output JSON or plain HTML. If this
definition is refactored, it can use the decorator as well.