Changes the BTN provider to support backlog searches
The new provider uses BTN's API to search for torrents in stead of using
the RSS. The regular "RSS" searches are emulated by an API request for
the torrents added since the last successful request.

Because the BTN API is accessed using a JSON-RPC interface, the library
jsonrpclib is added. The source of this is
https://github.com/joshmarshall/jsonrpclib, I copied it with only some
minor changes to fix import errors without having to change the library
paths used in Sick Beard. The version used is from 20th of August 2011.
I have also added the jsonrpclib to the list of libraries as listed in the
readme.md.

Some changes had to be made to the way whole seasons are treated because
torrents cannot be split like NZB's can. This has been solved by checking
if the provider of the season file is BTN, if so the whole season is
added as a multi-ep result.

Because BTN often uses the name "tvrip" to identify SDTV (non-scene) rips,
I have added that to the name checking regex.