Adds ability to set the SABnzbd/NZBGet download queue priority for each show.
Download queue priorities can be defined for both recent episodes and
older episodes. The available priorities are "Default" and "High".
The "Default" priority will use the priority configured in SABnzb for
the Category of that NZB, while in NZBGet the NZB will be added
to the bottom of the download queue. The "High" priority will cause the
NZB to be added to SABnzbd with a priority of "High", while in NZBGet
the NZB will be added to the top of the queue.

Sick-Beard defaults to assigning recent episodes a priority of
High, and older episodes a priority of Default. It is possible to
set different priorities when adding a new show, as well as change the
Sick-Beard defaults. The priorities for an existing show can be changed
within that show's edit page.

It is possible to have access to all of SABNzbd's priorities. In order
for these extra priorities to become available, the config setting
"sab_all_priorities" should be set to 1.

Priorities are valid only for shows being downloaded using
SABnzbd or NZBGet. They have no effect on torrent downloads or when
sending NZBs to a black hole. All options and info regarding priorities
will be hidden if neither SABnzbd or NZBGet is configured as the
download method.