Force line ending conversion on all files git considers text
Putting "* text=auto" in .gitattributes overrides the user's autocrlf option, so
regardless of what they have set, all text files will be converted to LF on
commit.

On checkout, git will convert all text files to the system's native line-endings
or to the one specified locally by the core.eol setting.

This is desireable because it will prevent mixed line endings, which make it
difficult for text editors, diff tools, and git to properly merge text,
regardless of whether the user has set their git configuration correctly or not.