Mostly done.
Reintroduce foreign key link between shows and episodes.  Because the tvdb_id is really the primary key for Shows I changed the behavior of save() to require it be set by hand.

Still need to clean up templates and a few other files.  May want to revisit database schema versioning to be a little more straight forward.