Almost new init.freebsd
Removed the use of daemon to daemonize Sickbeard and made more use of
the default functions of FreeBSD and Sickbeard.
Also removed the need to edit the init file. The file gets it's settings
directly from the sickbeard config file.
Changed the status and stop functions to be a bit more resilient when
faced with strange situations.