Added play file by clicking on the filename in displayShow
Gives the user the possibility to choose an executable and enable the
function.
Once enabled the user can click on a filename of an episode in the
displayShow page to open the executable with the filename as a
parameter.
It should be extended to give the user the possibility to customize the
parameters given to the exec for broader compatibility.
Build on and tested with win8 and VLC.