Adding DVD ordering to Sick-Beard
I originally had a good commit messsage here but months of this PR
becoming stagnant and repeatedly rebasing have removed it.

It adds the option to order by DVD using the TVDB API. This fixes a
whole bunch of shows like Firefly, Futurama, Batman The Animated Series
and many more.