Changed the REQUIRE line to comply with the FreeBSD handbook
The REQUIRE should be LOGIN, not DAEMON, per the FreeBSD
Porter's handbook, in the FreeBSD init script. This is because sickbeard runs as a non-root user and
isn't part of the base system. Removed sabnzbd from the REQUIRE as well
since sabnzbd is not required for sickbeard to start and so should not
be listed in the REQUIRE line