Added additional rc variables to the FreeBSD init script
Added rc variables to allow specification of the Sick Beard host, port,
web interface username, and web interface password in /etc/rc.conf so
that users don't have to modify the init script to change these values.
This was done because user's shouldn't be modifying the init script
since init scripts are typically installed by the package and are
overwritten with each update. This changes allows Sick Beard to be
safely packaged into a port when the day comes that it's ready for that.