Removed sickbeard_flags reference from command_args variable
Modified the FreeBSD init script by removing the "${sickbeard_flags}"
reference from command_args because, per the FreeBSD Porter's handbook,
the flags variable should never be part of command_args. This is because
if the flags variable is defined it is added to the command
automatically by rc(8) when the init script is run.

Unfortunately we currently can't define the flags variable because it
gets added as a flag to /usr/sbin/daemon, not SickBeard.py