Correct checking for TV Download Dir deletion
At the beginning of processTV.processDir, the path given in
dirName is expanded with os.path.realpath, which eliminates all
symbolic links. Later, when script checks whether current
post-processed folder should be deleted, it tries to match the
dirName against the sickbeard.TV_DOWNLOAD_DIR. However the
during that check, the second path is not expanded with
os.path.realpath which results in an unmatched condition and
hence deletion of the TV Download Dir if the path given in the
configuration is a symbolic link.