Better FreeBSD init script
A lot of things that were made in the init script are not needed
anymore. SickBeard closes cleanly when receiving a SIGTERM, so all the
wget part is obsolete.
The other cleanup (start, status, pid, etc.) has been made after
reading the sabnzbd script. Thus this init script is now very short and
easier to maintain I think.
It has been successfully tested on FreeBSD 10.0-RELEASE.