Support custom URIs for Transmission
Previously the user provided transmission address was trimmed, and only the
hostname and port was used when creating the transmissionrpc client, which uses
the default '/tranmission/rpc' path in this case. Now the scheme, port, and the
path is also passed when creating the client, which parses this information, and
looks for tranmission at the address specified by the user.