Save notification priorities as ints
pull request https://github.com/midgetspy/Sick-Beard/pull/886 started reading notifications as priorities but wasn't saving them as ints on notification updates. 

This caused the UI to always display the lowest priority in the select box, regardless of the actual values saved in the config, as there was a type mismatch, in addition subsequent saves would override the setting to be the lowest priority.