Fixed Unicode coercion errors in logging.
Some error messages in helpers.py were not correctly wrapping variables
in str() calls before attempting to append them to unicode output. This
caused unicode coercion errors which masked the actual errors triggering
the logging.

This change simply fixes the logging so it is more accurate.

There may be other instances of similar handling.  But I wanted to keep
this contained to the specific instances I was encountering in my
install.