fixed faststop behavior
rc.shutdown initiates a faststop command to the rc script which in sickbeard's case will not remove the PID file causing problems on the next start. I override the faststop command and run the normal stop command to fix this behaviour. Not sure if this is the best way to go about it, but it resolves the issue.